Revert testing changes for elevation
For some reason, elevation will not show in the layout designer preview in Android Studio but will show when actually running the device. Issue occurs on Buttons & TextViews so it is not specific to this library. Could not find a solid answer on what the problem was and many answers sounded like the shadow should be showing.

Thus, I will attribute it to an Android Studio bug or an issue on my machine but nevertheless, skipping since it seems irrelevant to the library.