Crude implementation of animation between positions
Am able to change position and animate nicely. It essentially ended up being the same way ceryle did it but I think my method is cleaner.

There was a bug with the bounce interpolator where it shows part of a button when going to the left.

Needs a major cleanup to look pretty. Might want to store the ValueAnimator so I can stop/cancel it at anytime. Also need to store the last position in case a user animates in another way.