Experimentation complete, able to add border and round the selected button
Actually, one minor thing I have to fix still. Not sure how I'm going to do that.

Currently, the border of the entire button group is on top of the selected button border. Major rework of library would be required to fix this and I don't think it's a necessary feature.